ExpFit

Documentation for ExpFit.

ExpFit.balanced_truncation_sub

ExpFit.esprit

ExpFit.esprit_sub

ExpFit.hankel_matrix

ExpFit.matrix_pencil

ExpFit.matrix_pencil_sub

ExpFit.balanced_truncation_sub — Method

balanced_truncation_sub(a::Vector{ComplexF64}, c::Vector{ComplexF64}, Um::AbstractMatrix{ComplexF64})

Sub function for balanced truncation.

ExpFit.esprit — Method

esprit(hk, dt, eps; ncols=nothing) -> exponent, coeff

Perform the ESPRIT algorithm using discrete data hk and the sampling interval dt.

ExpFit.esprit_sub — Method

esprit_sub(hk, eps; p=nothing)

Estimate the eigenvalues γ using the ESPRIT subspace method from the Hankel matrix constructed from hk.

ExpFit.hankel_matrix — Method

hankel_matrix(hk; q=nothing)

Construct a Hankel matrix from the discrete data hk (a vector of complex numbers).

ExpFit.matrix_pencil — Method

matrix_pencil_sub(hk, dt, eps; ncols) -> γ

Perform the Matrix Pencil method using discrete data hk and the sampling interval defined by [tmin, tmax].

ExpFit.matrix_pencil_sub — Method

matrix_pencil_sub(hk, eps; ncols) -> γ

Estimate the eigenvalues γ using the Matrix Pencil method from the Hankel matrix constructed from the discrete data hk.