Reference

ExpFit.balanced_truncation

ExpFit.balanced_truncation_sub

ExpFit.espira1

ExpFit.espira2

ExpFit.esprit

ExpFit.esprit_sub

ExpFit.fast_esprit

ExpFit.fast_esprit_sub

ExpFit.fnrm

ExpFit.hankel_matrix

ExpFit.matrix_pencil

ExpFit.matrix_pencil_sub

ExpFit.partial_lanczos_bidiagonalization

Inspection

ExpFit.balanced_truncation — Method

balanced_truncation(a::Vector{ComplexF64}, c::Vector{ComplexF64}, eps::Float64)

Given the exponents a and coefficients c of an exponential fitting, compute a new exponential fitting with a reduced number of terms.

ExpFit.balanced_truncation_sub — Method

balanced_truncation_sub(a::Vector{ComplexF64}, c::Vector{ComplexF64}, Um::AbstractMatrix{ComplexF64})

Sub function for balanced truncation.

ExpFit.espira1 — Method

espira1

This function first computes a modified FFT of y and places knots on the unit circle. It then calls the AAA routine from RationalFunctionApproximation.jl to obtain a rational approximant of (modified) DFT data, and finally extracts the ESPIRA parameters via a partial-fraction procedure.

ExpFit.espira2 — Method

espira2

This function computes a shifted FFT of its input data and uses AAA to obtain a rational approximant. It then extracts the exponential parameters by converting the resulting poles and residues into a sum of exponentials.

ExpFit.esprit — Method

esprit(hk, dt, eps; ncols=nothing) -> exponent, coeff

Perform the ESPRIT algorithm using discrete data hk and the sampling interval dt.

ExpFit.esprit_sub — Method

esprit_sub(hk, eps; p=nothing)

Estimate the eigenvalues γ using the ESPRIT subspace method from the Hankel matrix constructed from hk.

ExpFit.fast_esprit — Method

fast_esprit(hk, dt, eps; ncols=nothing) -> exponent, coeff

Perform the ESPRIT algorithm using discrete data hk and the sampling interval dt.

ExpFit.fast_esprit_sub — Method

fast_esprit_sub(hk, eps; p=nothing)

Estimate the eigenvalues γ using the ESPRIT subspace method from the Hankel matrix constructed from hk.

ExpFit.fnrm — Method

fnrm(s)

Generic weighted norm computation

ExpFit.hankel_matrix — Method

hankel_matrix(hk; q=nothing)

Construct a Hankel matrix from the discrete data hk (a vector of complex numbers).

ExpFit.matrix_pencil — Method

matrix_pencil_sub(hk, dt, eps; ncols) -> γ

Perform the Matrix Pencil method using discrete data hk and the sampling interval defined by [tmin, tmax].

ExpFit.matrix_pencil_sub — Method

matrix_pencil_sub(hk, eps; ncols) -> γ

Estimate the eigenvalues γ using the Matrix Pencil method from the Hankel matrix constructed from the discrete data hk.

ExpFit.partial_lanczos_bidiagonalization — Method

Generic Partial Lanczos Bidiagonalization

This function extends partiallanczosbidiagonalization_dble to handle both real and complex input vectors s.