Reference

ExpFit.Exponentials

ExpFit.balanced_truncation

ExpFit.balanced_truncation

ExpFit.espira1

ExpFit.espira1

ExpFit.espira1

ExpFit.espira1

ExpFit.espira1

ExpFit.espira1

ExpFit.espira2

ExpFit.espira2

ExpFit.espira2

ExpFit.espira2

ExpFit.espira2

ExpFit.espira2

ExpFit.esprit

ExpFit.esprit

ExpFit.esprit

ExpFit.esprit

ExpFit.esprit

ExpFit.esprit

ExpFit.expfit

ExpFit.expfit

ExpFit.expfit

ExpFit.expfit

ExpFit.expfit

ExpFit.expfit

ExpFit.expred

ExpFit.expred

ExpFit.fast_esprit

ExpFit.fast_esprit

ExpFit.fast_esprit

ExpFit.fast_esprit

ExpFit.fast_esprit

ExpFit.fast_esprit

ExpFit.matrix_pencil

ExpFit.matrix_pencil

ExpFit.matrix_pencil

ExpFit.matrix_pencil

ExpFit.matrix_pencil

ExpFit.matrix_pencil

ExpFit.prony

ExpFit.prony

ExpFit.prony

ExpFit.prony

ExpFit.prony

ExpFit.prony

Inspection

ExpFit.Exponentials — Type

Exponentials <: AbstractExpFit

A type representing a sum of exponentials of the form

f(t) = ∑ cᵢ exp(-aᵢ(t-t₀))

where aᵢ and cᵢ are the exponents and coefficients of the sum, respectively.

Fields

expon::AbstractVector{<:Number}: The exponents aᵢ of the sum of exponentials.

coeff::AbstractVector{<:Number}: The coefficients cᵢ of the sum of exponentials.

ExpFit.balanced_truncation — Method

balanced_truncation(a::AbstractVector{<:Number}, c::AbstractVector{<:Number}, eps::Float64) :: Exponentials

Given the exponents a and coefficients c of a sum of exponentials, compute a new sum of exponentials with a reduced number of terms for a given tolerance eps.

ExpFit.balanced_truncation — Method

balanced_truncation(a::AbstractVector{<:Number}, c::AbstractVector{<:Number}, M::Int) :: Exponentials

Given the exponents a and coefficients c of a sum of exponentials, compute a new sum of exponentials with a given number of terms M.

ExpFit.espira1 — Method

espira1(f::Vector{<:Number}, dt::Real, M::Int) :: Exponentials

Perform the ESPIRA-I algorithm using discrete data f and the sampling interval dt for a given model order M.

ExpFit.espira1 — Method

espira1(f::Vector{<:Number}, dt::Real, eps::Real) :: Exponentials

Perform the ESPIRA-I algorithm using discrete data f and the sampling interval dt for a given tolerance eps.

ExpFit.espira1 — Method

espira1(func::Function, tmin::Real, tmax::Real, nsamples::Int, M::Int) :: Exponentials

Perform the ESPIRA-I algorithm using a function func in the range [tmin,tmax] and nsamples sampling points for a given model order M.

ExpFit.espira1 — Method

espira1(func::Function, tmin::Real, tmax::Real, nsamples::Int, eps::Real) :: Exponentials

Perform the ESPIRA-I algorithm using a function func in the range [tmin,tmax] and nsamples sampling points for a given tolerance eps.

ExpFit.espira1 — Method

espira1(func::Function, tmin::Real, tmax::Real, dt::Real, M::Int) :: Exponentials

Perform the ESPIRA-I algorithm using a function func in the range [tmin,tmax] and a sampling interval dt for a given model order M.

ExpFit.espira1 — Method

espira1(func::Function, tmin::Real, tmax::Real, dt::Real, eps::Real) :: Exponentials

Perform the ESPIRA-I algorithm using a function func in the range [tmin,tmax] and a sampling interval dt for a given tolerance eps.

ExpFit.espira2 — Method

espira2(f::Vector{<:Number}, dt::Real, M::Int) :: Exponentials

Perform the ESPIRA-II algorithm using discrete data f and the sampling interval dt for a given model order M.

ExpFit.espira2 — Method

espira2(f::Vector{<:Number}, dt::Real, eps::Real)  :: Exponentials

Perform the ESPIRA-II algorithm using discrete data f and the sampling interval dt for a given tolerance eps.

ExpFit.espira2 — Method

espira2(func::Function, tmin::Real, tmax::Real, nsamples::Int, M::Int) :: Exponentials

Perform the ESPIRA-II algorithm using a function func in the range [tmin,tmax] and nsamples sampling points for a given model order M.

ExpFit.espira2 — Method

espira2(func::Function, tmin::Real, tmax::Real, nsamples::Int, eps::Real) :: Exponentials

Perform the ESPIRA-II algorithm using a function func in the range [tmin,tmax] and nsamples sampling points for a given tolerance eps.

ExpFit.espira2 — Method

espira2(func::Function, tmin::Real, tmax::Real, dt::Real, M::Int) :: Exponentials

Perform the ESPIRA-II algorithm using a function func in the range [tmin,tmax] and a sampling interval dt for a given model order M.

ExpFit.espira2 — Method

espira2(func::Function, tmin::Real, tmax::Real, dt::Real, eps::Real) :: Exponentials

Perform the ESPIRA-II algorithm using a function func in the range [tmin,tmax] and a sampling interval dt for a given tolerance eps.

ExpFit.esprit — Method

esprit(hk::Vector{<:Number}, dt::Real, M::Int; ncols::Union{Int,Nothing}=nothing) :: Exponentials

Perform the ESPRIT algorithm using discrete data hk and the sampling interval dt for a given model order M.

ExpFit.esprit — Method

esprit(hk::AbstractVector{<:Number}, dt::Real, eps::Real; ncols::Union{Int,Nothing}=nothing) :: Exponentials

Perform the ESPRIT algorithm using discrete data hk and the sampling interval dt for a given tolerance eps.

ExpFit.esprit — Method

esprit(func::Function, tmin::Real, tmax::Real, nsamples::Int, M::Int; ncols::Union{Int,Nothing}=nothing) :: Exponentials

Perform the ESPRIT algorithm using a function func in the range [tmin,tmax] and nsamples sampling points for a given model order M.

ExpFit.esprit — Method

esprit(func::Function, tmin::Real, tmax::Real, nsamples::Int, eps::Real; ncols::Union{Int,Nothing}=nothing) :: Exponentials

Perform the ESPRIT algorithm using a function func in the range [tmin,tmax] and nsamples sampling points for a given tolerance eps.

ExpFit.esprit — Method

esprit(func::Function, tmin::Real, tmax::Real, dt::Real, M::Int; ncols::Union{Int,Nothing}=nothing) :: Exponentials

Perform the ESPRIT algorithm using a function func in the range [tmin,tmax] and a sampling interval dt for a given model order M.

ExpFit.esprit — Method

esprit(func::Function, tmin::Real, tmax::Real, dt::Real, eps::Real; ncols::Union{Int,Nothing}=nothing) :: Exponentials

Perform the ESPRIT algorithm using a function func in the range [tmin,tmax] and a sampling interval dt for a given tolerance eps.

ExpFit.expfit — Method

expfit(f::AbstractVector{<:Number}, dt::Real, M::Int; alg::AbstractExpFit=ESPRIT()) :: Exponentials

Estimate the exponents and coefficients of the sum of exponentials for the function f using the algorithm alg.

Arguments

f::AbstractVector{<:Number}: The samples of the function.

dt::Real: The sampling interval.

M::Int: The model order.

alg::AbstractExpFit=ESPRIT(): The algorithm to use for the estimation.

ExpFit.expfit — Method

expfit(f::AbstractVector{<:Number}, dt::Real, eps::Real; alg::AbstractExpFit=ESPRIT()) :: Exponentials

Estimate the exponents and coefficients of the sum of exponentials for the function f using the algorithm alg.

Arguments

f::AbstractVector{<:Number}: The samples of the function.

dt::Real: The sampling interval.

eps::Real: The tolerance for the algorithm.

alg::AbstractExpFit=ESPRIT(): The algorithm to use for the estimation.

ExpFit.expfit — Method

expfit(f::AbstractVector{<:Number}, dt::Real, eps::Real; alg::AbstractExpFit=ESPRIT()) :: Exponentials

Estimate the exponents and coefficients of the sum of exponentials for the function f using the algorithm alg.

Arguments

f::AbstractVector{<:Number}: The samples of the function.

dt::Real: The sampling interval.

eps::Real: The tolerance for the algorithm.

alg::AbstractExpFit=ESPRIT(): The algorithm to use for the estimation.

ExpFit.expfit — Method

expfit(func::Function, tmin::Real, tmax::Real, nsamples::Int, eps::Real; alg::AbstractExpFit=ESPRIT()) :: Exponentials

Estimate the exponents and coefficients of the sum of exponentials for the function func using the algorithm alg.

Arguments

func::Function: The function to estimate the exponents and coefficients.

tmin::Real: The initial time.

tmax::Real: The final time.

nsamples::Int: The number of samples to take in the interval [tmin, tmax].

eps::Real: The tolerance for the algorithm.

alg::AbstractExpFit=ESPRIT(): The algorithm to use for the estimation.

ExpFit.expfit — Method

expfit(func::Function, tmin::Real, tmax::Real, dt::Real, M::Int; alg::AbstractExpFit=ESPRIT()) :: Exponentials

Estimate the exponents and coefficients of the sum of exponentials for the function func using the algorithm alg.

Arguments

func::Function: The function to estimate the exponents and coefficients.

tmin::Real: The initial time.

tmax::Real: The final time.

dt::Real: The sampling interval.

M::Int: The model order.

alg::AbstractExpFit=ESPRIT(): The algorithm to use for the estimation.

ExpFit.expfit — Method

expfit(func::Function, tmin::Real, tmax::Real, dt::Real, eps::Real; alg::AbstractExpFit=ESPRIT()) :: Exponentials

Estimate the exponents and coefficients of the sum of exponentials for the function func using the algorithm alg.

Arguments

func::Function: The function to estimate the exponents and coefficients.

tmin::Real: The initial time.

tmax::Real: The final time.

dt::Real: The sampling interval.

eps::Real: The tolerance for the algorithm.

alg::AbstractExpFit=ESPRIT(): The algorithm to use for the estimation.

ExpFit.expred — Method

expred(a::AbstractVector{<:Number}, c::AbstractVector{<:Number}, M::Int) :: Exponentials

Finding a new set of exponents and coefficients for the sum of exponentials using the balanced truncation method.

Arguments

a::AbstractVector{<:Number}: The exponents aᵢ of the sum of exponentials.

c::AbstractVector{<:Number}: The coefficients cᵢ of the sum of exponentials.

M::Int: The model order.

ExpFit.expred — Method

expred(a::AbstractVector{<:Number}, c::AbstractVector{<:Number}, eps::Real) :: Exponentials

Finding a new set of exponents and coefficients for the sum of exponentials using the balanced truncation method.

Arguments

a::AbstractVector{<:Number}: The exponents aᵢ of the sum of exponentials.

c::AbstractVector{<:Number}: The coefficients cᵢ of the sum of exponentials.

eps::Real: The tolerance for the balanced truncation method.

ExpFit.fast_esprit — Method

fast_esprit(hk::Vector{<:Number}, dt::Real, M::Int) :: Exponentials

Perform the Fast ESPRIT algorithm using discrete data hk and the sampling interval dt for a given model order M.

ExpFit.fast_esprit — Method

fast_esprit(hk::Vector{<:Number}, dt::Real, eps::Real) :: Exponentials

Perform the Fast ESPRIT algorithm using discrete data hk and the sampling interval dt for a given tolerance eps.

ExpFit.fast_esprit — Method

fast_esprit(func::Function, tmin::Real, tmax::Real, nsamples::Int, M::Int) :: Exponentials

Perform the Fast ESPRIT algorithm using a function func in the range [tmin,tmax] and nsamples sampling points for a given model order M.

ExpFit.fast_esprit — Method

fast_esprit(func::Function, tmin::Real, tmax::Real, nsamples::Int, eps::Real) :: Exponentials

Perform the Fast ESPRIT algorithm using a function func in the range [tmin,tmax] and nsamples sampling points for a given tolerance eps.

ExpFit.fast_esprit — Method

fast_esprit(func::Function, tmin::Real, tmax::Real, dt::Real, M::Int) :: Exponentials

Perform the Fast ESPRIT algorithm using a function func in the range [tmin,tmax] and a sampling interval dt for a given model order M.

ExpFit.fast_esprit — Method

fast_esprit(func::Function, tmin::Real, tmax::Real, dt::Real, eps::Real) :: Exponentials

Perform the Fast ESPRIT algorithm using a function func in the range [tmin,tmax] and a sampling interval dt for a given tolerance eps.

ExpFit.matrix_pencil — Method

matrix_pencil(hk::Vector{<:Number}, dt::Real, M::Int; ncols::Union{Int,Nothing}=nothing) :: Exponentials

Perform the matrix pencil algorithm using discrete data hk and the sampling interval dt for a given model order M.

ExpFit.matrix_pencil — Method

matrix_pencil(hk::AbstractVector{<:Number}, dt::Real, eps::Real; ncols::Union{Int,Nothing}=nothing) :: Exponentials

Perform the matrix pencil algorithm using discrete data hk and the sampling interval dt for a given tolerance eps.

ExpFit.matrix_pencil — Method

matrix_pencil(func::Function, tmin::Real, tmax::Real, nsamples::Int, M::Int; ncols::Union{Int,Nothing}=nothing) :: Exponentials

Perform the matrix pencil algorithm using a function func in the range [tmin,tmax] and nsamples sampling points for a given model order M.

ExpFit.matrix_pencil — Method

matrix_pencil(func::Function, tmin::Real, tmax::Real, nsamples::Int, eps::Real; ncols::Union{Int,Nothing}=nothing)

Perform the matrix pencil algorithm using a function func in the range [tmin,tmax] and nsamples sampling points for a given tolerance eps.

ExpFit.matrix_pencil — Method

matrix_pencil(func::Function, tmin::Real, tmax::Real, dt::Real, M::Int; ncols::Union{Int,Nothing}=nothing) :: Exponentials

Perform the matrix pencil algorithm using a function func in the range [tmin,tmax] and a sampling interval dt for a given model order M.

ExpFit.matrix_pencil — Method

matrix_pencil(func::Function, tmin::Real, tmax::Real, dt::Real, eps::Real; ncols::Union{Int,Nothing}=nothing) :: Exponentials

Perform the matrix pencil algorithm using a function func in the range [tmin,tmax] and a sampling interval dt for a given tolerance eps.

ExpFit.prony — Method

prony(hk::Vector{<:Number}, dt::Real, M::Int) :: Exponentials

Perform the Prony method using discrete data hk and the sampling interval dt for a given model order M.

ExpFit.prony — Method

prony(hk::Vector{<:Number}, dt::Real, eps::Real) :: Exponentials

Perform the Prony method using discrete data hk and the sampling interval dt for a given tolerance eps.

ExpFit.prony — Method

prony(func::Function, tmin::Real, tmax::Real, nsamples::Int, eps::Real) :: Exponentials

Perform the Prony method using a function func in the range [tmin,tmax] and nsamples sampling points for a given tolerance eps.

ExpFit.prony — Method

prony(func::Function, tmin::Real, tmax::Real, nsamples::Int, M::Int) :: Exponentials

Perform the Prony method using a function func in the range [tmin,tmax] and nsamples sampling points for a given model order M.

ExpFit.prony — Method

prony(func::Function, tmin::Real, tmax::Real, dt::Real, eps::Real) :: Exponentials

Perform the Prony method using a function func in the range [tmin,tmax] and a sampling interval dt for a given tolerance eps.

ExpFit.prony — Method

prony(func::Function, tmin::Real, tmax::Real, dt::Real, M::Int) :: Exponentials

Perform the Prony method using a function func in the range [tmin,tmax] and a sampling interval dt for a given model order M.